v0.1.5

+++ src/features/newfeed/components/Comments.tsx
@@ -1,9 +1,8 @@
 import { usePostDataMutation } from "@/redux/api/queryApi";
 import { Avatar } from "@mantine/core";
 import React, { Key } from "react";
-import toast from "react-hot-toast";
 import { BiTrash } from "react-icons/bi";
-
+import Swal from 'sweetalert2';
 interface CommentProps {
   data: {
     content: string;
@@ -21,19 +20,34 @@ const Comments: React.FC<CommentProps> = ({ data }) => {
   const [deleteComment] = usePostDataMutation();
 
   const deleteCommentHandler = async () => {
-    try {
-      const response = await deleteComment({
-        url: `/comments/${data?.id}`,
-        method: "DELETE",
-      });
-      // console.log(response);
-      if (response?.data?.status === "success") {
-        toast.success("Your comment deleted successfully!");
+    const willDelete = await Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    });
+
+    if (willDelete.isConfirmed) {
+      try {
+        const response = await deleteComment({
+          url: `/comments/${data?.id}`,
+          method: "DELETE",
+        });
+        console.log(response);
+        if (response?.data?.status === "success") {
+          Swal.fire("Deleted!", "Your comment has been deleted.", "success");
+          // Optionally reset state or fetch new comments here
+        }
+      } catch (error) {
+        console.error(error);
+        Swal.fire("Failed!", "Failed to delete the comment.", "error");
       }
-    } catch (error) {
-      console.log(error);
     }
   };
+
   return (
     <div className="flex w-full justify-between gap-3 items-center text-gray-500">
       <div className="flex gap-3 justify-between items-center ">
